Updated .gitmodules to get rid of untracked changes in snipmate plugin.
The bundle/snipmate.vim submodule was reporting it was modified by an
untracked change due to the bundle/snipmate.vim/doc/tags file generated
by :Helptags in vim. Creating .gitignore files in the submodule or the
main repo wouldn't fix this so 'ignore = dirt' was added to the snipmate
submodule in .gitmodules.