Stop formatting zig files on save
The autoformatting doesn't play nice with undo-redo; if you undo and
then save, the format operation replaces the whole buffer (even if the
formatting doesn't change anything), thus creating an undo point and
preventing the use of redo to go back. In addition (and more
importantly) the auto formatting causing responsivness issues for large
fileds that makes scrolling and undo-redo laggy (this is presumably some
interaction with either the lsp or auto-completion).